DynamoDBテーブル更新をバックエンドに反映。 APIGatewayからの呼び出しに対応。

+++ infra/cdk/lambda/delete-task/index.py
@@ -5,13 +5,17 @@ import boto3
 dynamo = boto3.resource('dynamodb')
 
 def handler(event, context):
+    user_id = event['requestContext']['authorizer']['claims']['sub']
     task_id = event['pathParameters']['taskId']
     table = dynamo.Table(os.environ['TASKS_TABLE_NAME'])
 
-    table.delete_item(Key={'taskId': task_id})
+    table.delete_item(Key={'userId': user_id, 'taskId': task_id})
 
     return {
         'statusCode': 204,
-        'headers': {'Content-Type': 'application/json'},
+        'headers': {
+            'Content-Type': 'application/json',
+            'Access-Control-Allow-Origin': '*'
+        },
         'body': ''
     }
